requirements.txt
# requirements.txt
torch
torchvision
Pillow
opencv-python
scikit-image
tensorflow
numpy
scikit-learn
